Added FoldersModel.
The FolderModel is responsible for retrieveing the folders displayed on the start page.

After an interesting discussion on the mailing-list, I decided to go on with the database view solution, rather than implementing all the fixed-items-dynamic-items stuff in C++.
It seemed easier to maintain.
So, thanks to the database 'folders' VIEW, FoldersModel is just a basic subclass of a QSqlQueryModel.
Time will tell us if it is the right approach or if we have to implement all this stuff in some other way, especially if performance is bad.